src/composite_handler/requirements.txt
asgiref==3.4.1
attrs==21.2.0
boto3==1.20.4
Django==3.2.9
django-environ==0.8.1
django-health-check==3.16.4
django-q==1.3.9
djangorestframework==3.12.4
jsonschema==4.2.1
pdf2image==1.16.0
psutil==5.8.0
psycopg2-binary==2.9.1
pyrsistent==0.18.0
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
tqdm==4.62.3
